requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
reportlab
